requirements.txt
customtkinter>=5.2.0
pystray>=0.19.0
Pillow>=10.0.0
pyperclip>=1.8.0
pynput>=1.7.6
python-dotenv>=1.0.0
requests>=2.31.0
google-genai>=1.0.0
groq>=0.4.0
ddgs>=6.0.0
pyinstaller>=6.0.0
